+++ CHANGELOG.md
@@ -7,7 +7,8 @@ All notable changes to Rainbow-CLI will be documented in this file.
 
 ### [1.36.0] - 2018-02-02
 - #CRRAINB-1518: Display host information when connected
-- #CRRAINB-1533: Block and unblock a user
+- #CRRAINB-1533: Add commands for blocking and unblocking a user
+- #CRRAINB-1601: Add command for creating a new application
 
 ### [1.35.0] - 2018-01-15
 - #RQRAINB-648: Add mass-provisionning commands